+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # insy
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @insy/server@0.0.8
+
 ## 0.0.2
 
 ### Patch Changes