+++ packages/client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @insy/client
 
+## 0.0.5
+
+### Patch Changes
+
+- set default model
+
 ## 0.0.4
 
 ### Patch Changes