+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @insy/server
 
+## 0.0.8
+
+### Patch Changes
+
+- set default model
+- Updated dependencies
+  - @insy/client@0.0.5
+
 ## 0.0.7
 
 ### Patch Changes