+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @insy/shared
 
+## 0.0.3
+
+### Patch Changes
+
+- set default model
+
 ## 0.0.2
 
 ### Patch Changes